#540 refactor(core): consolidate formatError into src/core/errors.ts across all private copies
Add a single canonical formatError (err instanceof Error ? err.message :
String(err)) to src/core/errors.ts and replace the 20 private copies across
src/ with imports from it. The richer cli/output.ts ('[<code>] <message>\n
hint: ...') and ui/src/lib/format-error.ts (null/object-aware) variants are
intentionally left untouched so their output stays stable.

src/runs/spawn/util.ts no longer exports formatError; its consumers
(plot-append.ts, seed-extensions.ts) now import from core/errors.ts directly.

Refs pl-4d2e step 10 (warren-bd4f).
