+++ src/core/errors.ts
@@ -30,3 +30,18 @@ export class ValidationError extends WarrenError {
 export class StateTransitionError extends WarrenError {
 	readonly code = "state_transition_error";
 }
+
+/**
+ * Canonical minimal error formatter: the `.message` for `Error` instances,
+ * otherwise `String(err)`. Used everywhere a reason string is slotted into a
+ * log field, a failure-envelope `reason`, or a wrapped error message.
+ *
+ * Deliberately code/hint-free — the richer `[<code>] <message>\n  hint: ...`
+ * rendering lives in `src/cli/output.ts` (terminal-facing) and the
+ * null/object-aware variant lives in `src/ui/src/lib/format-error.ts`
+ * (browser-facing); both are kept separate so their output stays stable.
+ */
+export function formatError(err: unknown): string {
+	if (err instanceof Error) return err.message;
+	return String(err);
+}
