+++ src/runs/spawn/util.ts
@@ -30,8 +30,3 @@ export async function resolveCoordinationProject(
 	}
 	return repos.projects.require(seedProjectId);
 }
-
-export function formatError(err: unknown): string {
-	if (err instanceof Error) return err.message;
-	return String(err);
-}
